--- src/main/java/com/mapper/mapperpractice/MapperPracticeApplication.java
@@ -1,13 +1,0 @@
-package com.mapper.mapperpractice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MapperPracticeApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MapperPracticeApplication.class, args);
-	}
-
-}
