+++ src/main/java/com/mapper/mapperpractice/entity/Name.java
@@ -1,4 +1,4 @@
-package com.mapper.mapperpractice;
+package com.mapper.mapperpractice.entity;
 
 public class Name {
     private  int id;
